net-snmp: update to 5.9.1

+++ release/src/router/net-snmp/agent/mibgroup/mibII/ip.c
@@ -883,7 +883,7 @@ ip_load(netsnmp_cache *cache, void *vmagic)
     int             i;
     static int      sname[4] = { CTL_NET, PF_INET, IPPROTO_IP, 0 };
     size_t          len;
-    int             magic = (int) vmagic;
+    int             magic = (uintptr_t) vmagic;
 
     switch (magic) {
     case IPFORWARDING:
